Add Approval Check for Payments
When calling the `/v2/payments` endpoint, the easiest thing to look at
in order to understand if the Payment attempt was succesful is
`result.approved`

Rather than requiring the calling code to work with `result` object
directly, we can add a basic `#approved?` predicate method on the
Payment class. This will allow the calling code to easily confirm if a
Payment was successful and act accordingly.