Extract Token Path to Class Method
The token path is the path used externally (in front-end
implementations) to make a request to tokenize a credit card. In
particular it is the path used by the VGS javascript library to post
credit card details.

It's best to expose this path via the gem so that client code has a
central and standard place to get the token path.

To make it easier for the client code, I also added a method on the Lpt
class, `Lpt.token_path` that delegates to the Instrument class