Improve Handling of Object Values in Resources
There are a few helper methods that have been added to the Payment and
Instrument resource classes that reach into objects (ie Hash) under the
hood. In cases where that object has not be initialized with a value,
the helper methods explode and raise a method not found error.

To better guard against these edge cases, I've introduced a check to
ensure that the object responds to `with_indifferent_access` before
attempting to pull a value