Optionally Include Protocol with API Base URL
The VGS javascript that is used on the front end for tokenizing credit
card info requires the base API url without the protocol. In order to
easily support this requirement, we can allow an optional param to be
passed into the `#api_base_url` method, which will return the base url
without the protocol.

To make this even easier, I added a delegator class method on the Lpt
class so that the client code can access `Lpt.api_base_url` directly.