Improve Error Handling for Create/Retrieve
We have a problem where the low level Faraday errors are bubbling up
outside of the context of this gem. It's not a great experience for gem
users.

We want to control the narrative around the errors at the gem level,
consuming the raw API errors from Faraday and then re-raise LPT owned
errors, which will allow the gem users to better understand what has
happened.

I've implemented this at a high level of granularity, right now just
resource creation failure and resource retrieval failure. These inherit
from a general LPT error. It's a very minimal implementation at the
moment but it sets us up to expand and scale the error handling over
time.

Gem users can get specifics about the error via the message attribute on
the error. If this proves to be too troublesome or gem users need more
granular error classes to manage their control flow.. we can cross that
bridge when we get there.